Small cleanup and presentation
