--- web/src/model/division/personnel_divided.ts
@@ -1,15 +1,0 @@
-export default class DividedPersonnel {
-    public id: number;
-    public firstName: string;
-    public lastName: string;
-
-    constructor(
-      id: number,
-      firstName: string = "",
-      lastName: string = "")
-    {
-      this.id = id;
-      this.firstName = firstName;
-      this.lastName = lastName;
-    }
-}
