+++ web/src/views/web/page/index.ejs
@@ -12,7 +12,10 @@
         <p>Welcome to the warehouse dashboard.</p>
 
         <a class="button" href="/personnel">Manage Personnel</a>
-        <a class="button" href="/divider">Personnel Division</a>
+
+        <h2 style="margin-top: 1em;">Utilities</h2>
+
+        <p><a class="button" href="/divider">Division - Personnel Who Work In Every Warehouse</a></p>
       </div>
     </main>
